refactor: better class naming
All classes are inside Rubocop namespace, adding 'Rubocop' to them is redundant.
