+++ src/rubocop/helpers/doc-cache.js
